(#51) Rely on Puppet 7 built-in windows utils
Prior to Puppet 7, the win32-process and win32-servic gems where bundled
in the Puppet AIO package for windows, but where removed as Puppet
gained support for these features in it's core:

- https://github.com/puppetlabs/puppet/commit/321f36fac7856fdd8af82f80d236fb2603a825ea
- https://github.com/puppetlabs/puppet/commit/50a908694916e812e53b5ab5b0a5302ed9ba9811

Update the code to use the new API when available, and fall-back to the
previous code.

This unbreaks the module on windows with Puppet 7.